adding some stuff to convert to 3d

+++ include/f2d/draw.h
@@ -1,9 +1,9 @@
 #ifndef DRAW_H
 #define DRAW_H
 
-#include <game/texture.h>
-#include <game/wm/window.h>
-#include <game/wm/wm.h>
+#include <f2d/texture.h>
+#include <f2d/wm/window.h>
+#include <f2d/wm/wm.h>
 
 #define DRAW_EVENT_SINGLE_DRAW 0
 #define DRAW_EVENT_CALLBACK 1
@@ -16,6 +16,6 @@ typedef struct {
 } draw_event_t;
 
 void draw_push_event(draw_event_t *event);
-void draw_all(display_t *display, window_t *window, unsigned shader_id);
+void draw_all(display_t *display, window_t *window, long shader_id);
 
 #endif
